ensuring tars don't get nested

+++ internet_of_fish/modules/utils/file_utils.py
@@ -21,6 +21,8 @@ def tar_directory(dir_path, new_name=None):
         new_name = dir_path / new_name
     with tarfile.open(new_name, 'w') as tarball:
         for f in dir_path.iterdir():
+            if f.suffix is '.tar':
+                continue
             tarball.add(f)
             f.unlink()
     return new_name
